Clean up test data,change file structure, rename some things
