+++ .gitignore
@@ -6,6 +6,9 @@
 
 output/
 acma/
+relACs/
+data/
+test_data/
 # Byte-compiled / optimized / DLL files
 __pycache__/
 *.py[cod]
