+++ src/MainPage.ui
@@ -2305,7 +2305,7 @@
      <x>0</x>
      <y>0</y>
      <width>1641</width>
-     <height>26</height>
+     <height>21</height>
     </rect>
    </property>
    <widget class="QMenu" name="menuSettings">
